fix, revision: docs, copypasta (#13)
* fix, revision: docs, copypasta

* fix: rustfmt

* fix: ci

* fix: doc test, clippy
