+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "rust-analyzer.cargo.features": [
+        "clipboard",
         "geolocation",
         "utils",
         "i18n"
