+++ Cargo.toml
@@ -16,7 +16,7 @@ categories = ["multimedia","os", "wasm"]
 #members = ["./examples/*"]
 
 [features]
-clipboard = ["dep:clipboard"]
+clipboard = ["dep:copypasta"]
 notifications = ["dep:notify-rust"]
 geolocation = ["dep:windows", "dep:futures-util", "dep:futures", "dep:web-sys", "dep:wasm-bindgen", "utils"]
 color_scheme = ["dep:web-sys", "dep:wasm-bindgen", "dep:wasm-bindgen-futures"]
@@ -36,7 +36,7 @@ cfg-if = "1.0.0"
 uuid = { version = "1.3.2", features = ["v4"], optional = true }
 async-broadcast = { version = "0.5.1", optional = true }
 # clipboard
-clipboard = { version = "0.5.0", optional = true }
+copypasta = { version = "0.8.2", optional = true }
 # notifications
 notify-rust = { version = "4.8.0", optional = true }
 # geolocation
@@ -59,5 +59,7 @@ js-sys = "0.3.62"
 uuid = {  version = "1.3.2", features = ["v4", "js"]}
 
 [package.metadata.docs.rs]
-all-features = true
+default-target = "x86_64-pc-windows-msvc"
+no-default-features = true
+features = ["desktop-testing"]
